refactor: Reflected changes on module naming
